fix(handle-doer): count affected rows
The method `mimer_driver::pdo_mimer_handle_doer()` should only return number of affected rows upon a query using `UPDATE`, `INSERT`, or `DELETE`. Thus, a query with a result set and should not have affected any rows, or the statement was a DDL statement. The method should return a zero-value, which the method already correctly does.

Resolves #9.