feat(LOB): Big overhaul on LOB insertion handling, including updated tests
MimerSetXData functions require a buffer with only valid data (i.e. no partial multi-byte character sequences). This was a problem since we are reading a stream in chunks, so there had to be a way to validate that each chunk only contained valid MB characters. This commit (which yes, is way too big), refactores some of the old code as well as provide the mechanisms needed to get valid chunks of character data. I also updated the tests to (amongst other things) verify the inserted number of characters/bytes.