fix: adapted reference unwrapping to work for LOBs
Previous implementation did not work for LOBs, or rather file pointer parameters. I'm not 100% certain but I suspect a zval copy was being made, and that the new copy of the resource was not open (in any case it wasn't seekable, which caused the lob length checking to fail). This fix restored the funcitonality.